Legal-chatbot/requirements.txt
Flask               # if you use FastAPI instead of Flask           # to run FastAPI apps
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community   
pydantic>=2.0.0
pytest>=7.0.0
cohere
streamlit